requirements.txt
discord.py>=2.3.0
rembg>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
onnxruntime
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.9.0
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0